refactor id to Long

+++ src/main/java/com/sportgames/dao/PlaygroundDAO.java
@@ -7,11 +7,11 @@ import java.util.List;
 
 public interface PlaygroundDAO {
 
-    Playground findById(int id);
+    Playground findById(Long id);
     List<Playground> getAll();
     List<Playground> getPlaygroundBySportType(String type);
     void add(Playground playground);
-    void delete(int id);
-    Playground get(int id);
+    void delete(Long id);
+    Playground get(Long id);
     Playground findByName(String name);
 }
