+++ src/main/java/com/sportgames/dao/SportDAO.java
@@ -5,10 +5,10 @@ import com.sportgames.model.Sport;
 import java.util.List;
 
 public interface SportDAO {
-    Sport findById(int id);
+    Sport findById(Long id);
     List<Sport> getAll();
     void add(Sport sport);
-    void delete(int id);
-    Sport get(int id);
+    void delete(Long id);
+    Sport get(Long id);
     Sport findByName(String name);
 }
